+++ src/index.js
@@ -1,3 +1,11 @@
 module.exports = {
-    Client: require('./objects/Client')
+    Client: require('./objects/Client'),
+    Chat: require('./objects/Chat'),
+    ChatUser: require('./objects/ChatUser'),
+    Comment: require('./objects/Comment'),
+    Invite: require('./objects/Invite'),
+    Message: require('./objects/Message'),
+    Post: require('./objects/Post'),
+    User: require('./objects/User')
+
 }